moved Feature Store record to commons
