+++ litti-inference-router/internal/feature/featureStore.go
@@ -5,6 +5,8 @@ import (
 	"encoding/json"
 	"errors"
 
+	"com/litti/ml/litti-inference-router/internal/dto"
+
 	"github.com/redis/go-redis/v9"
 )
 
@@ -14,11 +16,11 @@ var redisClient = redis.NewClient(&redis.Options{
 	DB:       0,  // use default DB
 })
 
-func FetchFeatureGroupRow(featureGroupKey string) (map[string]FeatureStoreRecord, error) {
+func FetchFeatureGroupRow(featureGroupKey string) (map[string]dto.FeatureStoreRecord, error) {
 	return fetchRedisFeatureGroup(featureGroupKey)
 }
 
-func fetchRedisFeatureGroup(featureGroupKey string) (map[string]FeatureStoreRecord, error) {
+func fetchRedisFeatureGroup(featureGroupKey string) (map[string]dto.FeatureStoreRecord, error) {
 	ctx := context.Background()
 	val, err := redisClient.Get(ctx, featureGroupKey).Result()
 	if err != nil {
@@ -27,7 +29,7 @@ func fetchRedisFeatureGroup(featureGroupKey string) (map[string]FeatureStoreReco
 		}
 		return nil, err
 	}
-	var records map[string]FeatureStoreRecord
+	var records map[string]dto.FeatureStoreRecord
 	err = json.Unmarshal([]byte(val), &records)
 	if err != nil {
 		return nil, err
